Fix compilation with and without bundled frontend
